+++ src/modules/catalog/endpoints/create-product.endpoint.ts
@@ -5,6 +5,7 @@ import {
   HttpStatus,
   Group,
 } from '../../../../lib';
+import { permissions } from '../permissions';
 import { CreateProductDto } from '../dto/create-product.dto';
 import { Product } from '../entities/product.entity';
 
@@ -15,7 +16,7 @@ export class CreateProductEndpoint extends Endpoint<null, CreateProductDto> {
   private readonly products = this.db.getRepository(Product);
 
   async main() {
-    this.auth.assert('catalog.products.manage');
+    this.auth.assert(permissions['products.manage']);
 
     const product = await this.products.save(
       this.products.create({ name: this.body.name, stock: this.body.stock }),