+++ src/modules/identity/endpoints/create-user.endpoint.ts
@@ -6,6 +6,7 @@ import {
   HttpStatus,
   Group,
 } from '../../../../lib';
+import { permissions } from '../permissions';
 import { CreateUserDto } from '../dto/create-user.dto';
 import { User } from '../entities/user.entity';
 import { hashPassword } from '../services/password';
@@ -17,7 +18,7 @@ export class CreateUserEndpoint extends Endpoint<null, CreateUserDto> {
   private readonly users = this.db.getRepository(User);
 
   async main() {
-    this.auth.assert('identity.users.manage');
+    this.auth.assert(permissions['users.manage']);
 
     const taken = await this.users.existsBy({ username: this.body.username });
     if (taken) {